// Front End/livreur/src/app/core/models/user.model.ts
export class User {
    id: number;
    nom: string;
    prenom: string;
    dateNaissance: string;
    Tel: string;
    email: string;
    password: string;
    codePostal: String;
    image: String;
    pays: String;
    ville: String;
    username: string;
    state: string;
    roles: string[];
    accessToken: string;
}